+++ app/Icons/page.tsx
@@ -1,9 +1,8 @@
-import Link from "next/link";
 import { Container } from "react-bootstrap";
 import IconHeader from "../WebsiteHeaders/IconsHeaders/icons_header";
 import IconsDescription from "../WebsiteHeaders/IconsHeaders/icons_description";
 import IconsTutorial from "../WebsiteHeaders/IconsHeaders/icons_tutorial";
-import PaymentSuccess from "../WebsiteComponents/PaymentForm/PaymentSuccess";
+import PaymentSuccess from "../WebsiteComponents/IconsComponents/PaymentForm/PaymentSuccess";
 import ScrollingIcons from "../WebsiteComponents/IconsComponents/scrolling_icons";
 import IconsPayment from "../WebsiteComponents/IconsComponents/icons_payment";
