+++ app/WebsiteComponents/HomeComponents/Proj/projects.tsx
@@ -1,6 +1,6 @@
 import { Container, Row } from "react-bootstrap";
 import ProjectComponent from "./project_component";
-import { projects } from "../../../public/content/projects_content";
+import { projects } from "../../../../public/content/projects_content";
 
 export default function Projects() {
     return (
