+++ app/WebsiteComponents/HomeComponents/Research/research_component.tsx
